+++ relay-cardinality/src/redis/quota.rs
@@ -1,3 +1,7 @@
+use hash32::Hasher;
+use std::fmt;
+use std::hash::Hash;
+
 use relay_base_schema::metrics::MetricNamespace;
 use relay_base_schema::project::ProjectId;
 use relay_common::time::UnixTimestamp;
@@ -8,31 +12,38 @@ use crate::window::Slot;
 use crate::{CardinalityLimit, CardinalityScope, OrganizationId, Scoping, SlidingWindow};
 
 /// A quota scoping extracted from a [`CardinalityLimit`] and a [`Scoping`].
-#[derive(Clone, Copy, Debug, Hash, PartialEq, Eq)]
-pub struct QuotaScoping {
+///
+/// The partial quota scoping can be used to select/match on cardinality entries
+/// but it needs to be completed into a [`QuotaScoping`] by using
+/// [`PartialQuotaScoping::complete`].
+#[derive(Clone, Copy, Debug, Hash, PartialEq, Eq, PartialOrd, Ord)]
+pub struct PartialQuotaScoping {
     pub namespace: Option<MetricNamespace>,
     pub organization_id: Option<OrganizationId>,
     pub project_id: Option<ProjectId>,
     window: SlidingWindow,
+    scope: CardinalityScope,
 }
 
-impl QuotaScoping {
-    /// Creates a new [`QuotaScoping`] from a [`Scoping`] and [`CardinalityLimit`].
+impl PartialQuotaScoping {
+    /// Creates a new [`PartialQuotaScoping`] from a [`Scoping`] and [`CardinalityLimit`].
     ///
     /// Returns `None` for limits with scope [`CardinalityScope::Unknown`].
     pub fn new(scoping: Scoping, limit: &CardinalityLimit) -> Option<Self> {
         let (organization_id, project_id) = match limit.scope {
             CardinalityScope::Organization => (Some(scoping.organization_id), None),
             CardinalityScope::Project => (Some(scoping.organization_id), Some(scoping.project_id)),
+            CardinalityScope::Name => (Some(scoping.organization_id), Some(scoping.project_id)),
             // Invalid/unknown scope -> ignore the limit.
             CardinalityScope::Unknown => return None,
         };
 
         Some(Self {
-            window: limit.window,
             namespace: limit.namespace,
             organization_id,
             project_id,
+            window: limit.window,
+            scope: limit.scope,
         })
     }
 
@@ -64,6 +75,32 @@ impl QuotaScoping {
         UnixTimestamp::from_secs(timestamp.as_secs() + shift)
     }
 
+    /// Creates a [`QuotaScoping`] from the partial scoping and the passed [`Entry`].
+    ///
+    /// This unconditionally creates a quota scoping from the passed entry and
+    /// does not check whether the scoping even applies to the entry. The caller
+    /// needs to ensure this by calling [`Self::matches`] prior to calling `complete`.
+    pub fn complete(self, entry: Entry<'_>) -> QuotaScoping {
+        let name = match self.scope {
+            CardinalityScope::Name => Some(fnv32(entry.name)),
+            _ => None,
+        };
+
+        QuotaScoping { parent: self, name }
+    }
+}
+
+/// A quota scoping extracted from a [`CardinalityLimit`], a [`Scoping`]
+/// and completed with a [`CardinalityItem`](crate::CardinalityItem).
+///
+/// The scoping must be created using [`PartialQuotaScoping::complete`].
+#[derive(Clone, Copy, Debug, Hash, PartialEq, Eq, PartialOrd, Ord)]
+pub struct QuotaScoping {
+    parent: PartialQuotaScoping,
+    name: Option<u32>,
+}
+
+impl QuotaScoping {
     /// Returns the minimum TTL for a Redis key created by [`Self::into_redis_key`].
     pub fn redis_key_ttl(&self) -> u64 {
         self.window.window_seconds
@@ -74,7 +111,34 @@ impl QuotaScoping {
         let organization_id = self.organization_id.unwrap_or(0);
         let project_id = self.project_id.map(|p| p.value()).unwrap_or(0);
         let namespace = self.namespace.map(|ns| ns.as_str()).unwrap_or("");
+        let name = DisplayOptMinus(self.name);
 
-        format!("{KEY_PREFIX}:{KEY_VERSION}:scope-{{{organization_id}-{project_id}-{namespace}}}-{slot}")
+        format!("{KEY_PREFIX}:{KEY_VERSION}:scope-{{{organization_id}-{project_id}-{namespace}}}-{name}{slot}")
     }
+}
+
+impl std::ops::Deref for QuotaScoping {
+    type Target = PartialQuotaScoping;
+
+    fn deref(&self) -> &Self::Target {
+        &self.parent
+    }
+}
+
+struct DisplayOptMinus<T>(Option<T>);
+
+impl<T: fmt::Display> fmt::Display for DisplayOptMinus<T> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        if let Some(inner) = self.0.as_ref() {
+            write!(f, "{inner}-")
+        } else {
+            Ok(())
+        }
+    }
+}
+
+fn fnv32(s: &str) -> u32 {
+    let mut hasher = hash32::FnvHasher::default();
+    s.hash(&mut hasher);
+    hasher.finish32()
 }